fix: show slide content based on image or video
